Fix fetch from github
The fetch=".." directive is relative to the URL repository hosting the
manifest, as without this fix, a repository like android_build would be
fetched from https://git.replicant.us/LineageOS/android_build.git.

The manifest.xml doesn't have any reference to git.replicant.us.

The only reference to git.replicant.us is from the URL of the manifest
repository which in this case is is at the following URL:
https://git.replicant.us/GNUtoo/manifest.git.

This behavior is however not documented in the repo manifest format[1].

References:
-----------
[1]https://android.googlesource.com/tools/repo/+/refs/tags/v1.13.2/docs/repo-hooks.md

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>