Add Replicant repositories
At first we tried to move LineageOS's default.xml in a
subdirectory to keep as little changes as possible with LineageOS
in that file. However git didn't properly detect that the file
was moved as some changes in the files were necessary.

As the Replicant includes are now at the end of the file, it
should minimize the non-automatic rebases.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Rebased the commit and adapted it for LineageOS 17.1
Signed-off-by: Joonas Kylmälä <joonas.kylmala@iki.fi>