Add libsamsung-ipc for testing
At the time of writing, libsamsung-ipc is still not ready
for Replicant 9 as there is still ongoing work to cleanup
and integrate the patches that makes the modem work with
the replicant 9 kernel driver.

However since this integration work is going on, we also
need to make sure that libsamsung-ipc also compiles on
Replicant 9.

This change will ensure that libsamsung-ipc is built and
shipped inside the image, so if the compilation breaks
at some point, it will enable to detect it more easily.

The kernel driver and libsamsung-ril also need to be
cleaned up, and the ability to launch libsamsung-ipc
twice without crashing the device needs to be implemented.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>